requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
pypdf
python-dotenv
openai
tiktoken
pysqlite3-binary
